Move TCP Client logic to its package
