+++ pkg/logs/client/tcp/connection_manager_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
 
-package client
+package tcp
 
 import (
 	"net"
@@ -23,7 +23,7 @@ func newConnectionManagerForAddr(addr net.Addr) *ConnectionManager {
 }
 
 func newConnectionManagerForHostPort(host string, port int) *ConnectionManager {
-	endpoint := Endpoint{Host: host, Port: port}
+	endpoint := config.Endpoint{Host: host, Port: port}
 	return NewConnectionManager(endpoint)
 }
