+++ pkg/logs/client/tcp/delimiter_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
 
-package client
+package tcp
 
 import (
 	"testing"
