+++ pkg/logs/client/tcp/test_utils.go
@@ -3,9 +3,10 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
 
-package client
+package tcp
 
 import (
+	"github.com/DataDog/datadog-agent/pkg/logs/config"
 	"net"
 )
 
@@ -21,9 +22,9 @@ func AddrToHostPort(remoteAddr net.Addr) (string, int) {
 }
 
 // AddrToEndPoint creates an EndPoint from an Addr.
-func AddrToEndPoint(addr net.Addr) Endpoint {
+func AddrToEndPoint(addr net.Addr) config.Endpoint {
 	host, port := AddrToHostPort(addr)
-	return Endpoint{Host: host, Port: port}
+	return config.Endpoint{Host: host, Port: port}
 }
 
 // AddrToDestination creates a Destination from an Addr
