+++ pkg/logs/config/endpoints.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
 
-package client
+package config
 
 // Endpoint holds all the organization and network parameters to send logs to Datadog.
 type Endpoint struct {
