+++ pkg/logs/pipeline/provider.go
@@ -6,10 +6,10 @@
 package pipeline
 
 import (
+	"github.com/DataDog/datadog-agent/pkg/logs/client/tcp"
 	"sync/atomic"
 
 	"github.com/DataDog/datadog-agent/pkg/logs/auditor"
-	"github.com/DataDog/datadog-agent/pkg/logs/client"
 	"github.com/DataDog/datadog-agent/pkg/logs/config"
 	"github.com/DataDog/datadog-agent/pkg/logs/message"
 	"github.com/DataDog/datadog-agent/pkg/logs/restart"
@@ -28,15 +28,15 @@ type provider struct {
 	auditor           *auditor.Auditor
 	outputChan        chan *message.Message
 	processingRules   []*config.ProcessingRule
-	endpoints         *client.Endpoints
+	endpoints         *config.Endpoints
 
 	pipelines            []*Pipeline
 	currentPipelineIndex int32
-	destinationsContext  *client.DestinationsContext
+	destinationsContext  *tcp.DestinationsContext
 }
 
 // NewProvider returns a new Provider
-func NewProvider(numberOfPipelines int, auditor *auditor.Auditor, processingRules []*config.ProcessingRule, endpoints *client.Endpoints, destinationsContext *client.DestinationsContext) Provider {
+func NewProvider(numberOfPipelines int, auditor *auditor.Auditor, processingRules []*config.ProcessingRule, endpoints *config.Endpoints, destinationsContext *tcp.DestinationsContext) Provider {
 	return &provider{
 		numberOfPipelines:   numberOfPipelines,
 		auditor:             auditor,
