+++ pkg/logs/sender/sender_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/DataDog/datadog-agent/pkg/logs/client/mock"
+	"github.com/DataDog/datadog-agent/pkg/logs/client/tcp"
 	"github.com/DataDog/datadog-agent/pkg/logs/config"
 	"github.com/DataDog/datadog-agent/pkg/logs/message"
 )
@@ -29,10 +30,10 @@ func TestSender(t *testing.T) {
 	input := make(chan *message.Message, 1)
 	output := make(chan *message.Message, 1)
 
-	destinationsCtx := client.NewDestinationsContext()
+	destinationsCtx := tcp.NewDestinationsContext()
 	destinationsCtx.Start()
 
-	destination := client.AddrToDestination(l.Addr(), destinationsCtx)
+	destination := tcp.AddrToDestination(l.Addr(), destinationsCtx)
 	destinations := client.NewDestinations(destination, nil)
 
 	sender := NewSender(input, output, destinations)
@@ -60,13 +61,13 @@ func TestSenderNotBlockedByAdditional(t *testing.T) {
 	input := make(chan *message.Message, 1)
 	output := make(chan *message.Message, 1)
 
-	destinationsCtx := client.NewDestinationsContext()
+	destinationsCtx := tcp.NewDestinationsContext()
 	destinationsCtx.Start()
 
-	mainDestination := client.AddrToDestination(l.Addr(), destinationsCtx)
+	mainDestination := tcp.AddrToDestination(l.Addr(), destinationsCtx)
 	// This destination doesn't exists
-	additionalDestination := client.NewDestination(client.Endpoint{Host: "dont.exist.local", Port: 0}, destinationsCtx)
-	destinations := client.NewDestinations(mainDestination, []*client.Destination{additionalDestination})
+	additionalDestination := tcp.NewDestination(config.Endpoint{Host: "dont.exist.local", Port: 0}, destinationsCtx)
+	destinations := client.NewDestinations(mainDestination, []client.Destination{additionalDestination})
 
 	sender := NewSender(input, output, destinations)
 	sender.Start()
